Refactor: extract apply_spec_to_config, split experiment_handler, add ExecuteResult
- Extract apply_training_overrides() to tuner/core/config.py (Minor 1)
- CloudPipelineHandler: make shared methods public, remove private coupling (Minor 2)
- Fix sanitize_hf_job_labels: encode/decode '/' in label values (Minor 4)
- Add CloudEvalResult protocol to tuner/core/interfaces.py (Minor 5)
- Split experiment_handler.py (876L) into hf_training/eval/loss_stage_runner.py (Minor 3)
- Add ExecuteResult dataclass; HFJobsBackend.execute() returns structured result (Future 2)
- Fix stale _resolve_display_scenarios test reference
