+++ docs/benchmarks/model_hardware_benchmark_ledger.csv
@@ -1,4 +1,13 @@
 recorded_at,experiment_id,experiment_name,benchmark_group,base_model,model_size_bucket,dataset_variant,method,epochs,max_seq_length,train_flavor,eval_flavor,loss_flavor,train_batch_size,train_grad_accum,train_effective_batch,train_time_seconds,train_cost_usd_est,eval_time_seconds,eval_cost_usd_est,loss_time_seconds,loss_cost_usd_est,total_time_seconds,total_cost_usd_est,training_final_loss,eval_passed,eval_failed,eval_warned,eval_total,eval_pass_rate,eval_schema_pass_rate,loss_examples,loss_mean,loss_status,notes
-2026-03-22T20:00:00+00:00,exp_manual_l40_round1,qwen3_4b_l40_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,l40sx1,l40sx1,unavailable,14,3,42,3701.2,1.851,1631.735,0.816,,,5332.935,2.666,0.5291,45,22,10,77,0.5844155844,0.7142857143,,,failed_missing_peft,"Training was near OOM on L40S; eval quality tied with A100-scale runs."
-2026-03-22T20:00:00+00:00,exp_manual_a100_round1,qwen3_4b_a100_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,a100-large,a100-large,unavailable,27,2,54,2901.1,2.015,1607.75,1.116,,,4508.85,3.131,0.5595,44,22,11,77,0.5714285714,0.7142857143,,,failed_missing_peft,"Training was materially faster and much healthier on VRAM headroom than L40S."
-2026-03-22T20:00:00+00:00,exp_manual_a100x4_eval_round1,qwen3_4b_a100x4_eval_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,a100-large,a100x4,unavailable,27,2,54,2901.1,2.015,1597.624,4.438,,,4498.724,6.452,0.5595,45,22,10,77,0.5844155844,0.7142857143,,,failed_missing_peft,"Eval generation fell back to single-GPU because the base model is prequantized bnb-4bit; x4 did not yield a valid loss throughput comparison."
+2026-03-22T20:00:00+00:00,exp_manual_l40_round1,qwen3_4b_l40_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,l40sx1,l40sx1,unavailable,14,3,42,3701.2,1.851,1631.735,0.816,,,5332.935,2.666,0.5291,45,22,10,77,0.5844155844,0.7142857143,,,failed_missing_peft,Training was near OOM on L40S; eval quality tied with A100-scale runs.
+2026-03-22T20:00:00+00:00,exp_manual_a100_round1,qwen3_4b_a100_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,a100-large,a100-large,unavailable,27,2,54,2901.1,2.015,1607.75,1.116,,,4508.85,3.131,0.5595,44,22,11,77,0.5714285714,0.7142857143,,,failed_missing_peft,Training was materially faster and much healthier on VRAM headroom than L40S.
+2026-03-22T20:00:00+00:00,exp_manual_a100x4_eval_round1,qwen3_4b_a100x4_eval_round1,qwen3_4b_pruned_speed_round_1,Qwen/Qwen3-4B,4b,pruned_result_echo_linesdelta_recommendations,sft,2,2048,a100-large,a100x4,unavailable,27,2,54,2901.1,2.015,1597.624,4.438,,,4498.724,6.452,0.5595,45,22,10,77,0.5844155844,0.7142857143,,,failed_missing_peft,Eval generation fell back to single-GPU because the base model is prequantized bnb-4bit; x4 did not yield a valid loss throughput comparison.
+2026-03-23T10:47:24.940044+00:00,exp_20260323_104724_8e59,smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,0.000564,,0.000572,,0.000539,,0.001675,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:47:25.066626+00:00,exp_20260323_104724_a4a9,resume-smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,,,0.000599,,0.000668,,0.001267,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:47:25.156061+00:00,exp_20260323_104725_2ae8,resume-loss,train_eval_loss_smoke,Qwen/Qwen3-4B,4b,sample,sft,,,,,,,,,,,,,0.000582,,0.000582,,,,,,,,,1,0.2,completed,
+2026-03-23T10:48:15.883835+00:00,exp_20260323_104815_9f2f,smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,0.000535,,0.000517,,0.000499,,0.0015509999999999999,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:48:16.072846+00:00,exp_20260323_104815_5292,resume-smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,,,0.000685,,0.000648,,0.001333,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:48:16.165146+00:00,exp_20260323_104816_cd30,resume-loss,train_eval_loss_smoke,Qwen/Qwen3-4B,4b,sample,sft,,,,,,,,,,,,,0.000798,,0.000798,,,,,,,,,1,0.2,completed,
+2026-03-23T10:49:26.033205+00:00,exp_20260323_104925_eb49,smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,0.000544,,0.000509,,0.000499,,0.001552,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:49:26.224844+00:00,exp_20260323_104926_94c8,resume-smoke,train_eval_loss_smoke,HuggingFaceTB/SmolLM2-1.7B-Instruct,1.7b,sample,sft,,,,,,,,,,,0.000635,,0.000598,,0.0012330000000000002,,,1,0,0,1,,,1,0.4,completed,
+2026-03-23T10:49:26.317963+00:00,exp_20260323_104926_234d,resume-loss,train_eval_loss_smoke,Qwen/Qwen3-4B,4b,sample,sft,,,,,,,,,,,,,0.000894,,0.000894,,,,,,,,,1,0.2,completed,
