+++ tuner/cloud/__init__.py
@@ -9,6 +9,7 @@ from .hf_jobs import (
     build_hf_job_secrets,
     build_secrets_from_env,
     build_repo_checkout_steps,
+    decode_hf_job_label,
     format_timeout_hours,
     load_huggingface_hub,
     resolve_hf_bucket_id,
@@ -24,6 +25,7 @@ __all__ = [
     "build_hf_job_secrets",
     "build_secrets_from_env",
     "build_repo_checkout_steps",
+    "decode_hf_job_label",
     "format_timeout_hours",
     "load_huggingface_hub",
     "resolve_hf_bucket_id",
